It works
It reads and writes to the same Data0 and Data1 lines.  However because
the the reader holds the data0 and data1 lines high a direct splice into
the lines is NOT possible.  So Next I need to read in on one set of pins
and write out on another.  so in effect every time somone badges in it
wil read, write to the Filesystem then write out to the badging
infrastructure.